requirements.txt
altair==5.2.0
argcomplete==3.2.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
decorator==5.1.1
distlib==0.3.8
duckdb==0.10.0
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
idna==3.6
installer==0.7.0
ipython==8.22.2
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pydeck==0.8.1b0
Pygments==2.17.2
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==3.6.2
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
streamlit==1.32.1
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trove-classifiers==2024.3.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
userpath==1.9.2
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
